Assignment3 submission
Will commit folder re-arrangement later

+++ src/apis/DataBaseHandler.java
@@ -6,9 +6,9 @@ public interface DataBaseHandler{
     //store int l into something? maybe a file
     //send message to successfully store data
 
-    public long sendData(String s);
+    public int sendData(String s);
     //accepts the location of a file to process
 
-
+    
 
 }
